TransportManager split into partials
