+++ Core/Transport/TransportManager.Participant.cs
@@ -1,0 +1,16 @@
+namespace GodotPeer2PeerSteamCSharp.Core;
+
+public partial class TransportManager
+{
+    public bool IsConnectionActive()
+    {
+        return _transportService.IsConnectionActive();
+    }
+    
+    public void Disconnect()
+    {
+        _transportService.Disconnect();
+        ExecuteProcessMethodStatus(false);
+        Logger.Network("Disconnected from server.");
+    }
+}
