+++ Multiplayer/Services/Steam/Lobby/LobbyService.Host.cs
@@ -7,6 +7,11 @@ namespace GodotPeer2PeerSteamCSharp.Services.Steam.Lobby;
 
 public partial class LobbyService : ILobbyService
 {
+    private static void RegisterHostCallbacks()
+    {
+        SteamMatchmaking.OnLobbyCreated += OnLobbyCreatedCallback;
+    }
+    
     public async Task CreateLobby(int maxPlayers)
     {
         var lobby = await SteamMatchmaking.CreateLobbyAsync(maxPlayers);
@@ -16,11 +21,6 @@ public partial class LobbyService : ILobbyService
         _lobbyId = lobby.Value.Id;
         lobby.Value.SetPrivate();
         lobby.Value.SetJoinable(true);
-    }
-
-    private static void RegisterHostCallbacks()
-    {
-        SteamMatchmaking.OnLobbyCreated += OnLobbyCreatedCallback;
     }
 
     private static void OnLobbyCreatedCallback(Result result, Steamworks.Data.Lobby lobby)
